Transaction Edit form Improvement
