+++ src/components/Forms/Transactions/TransactionForm.tsx
@@ -11,8 +11,8 @@ import NoteIcon from "@/assets/icons/note.svg?icon";
 import Accordion from '@mui/material/Accordion';
 import AccordionSummary from '@mui/material/AccordionSummary';
 import AccordionDetails from '@mui/material/AccordionDetails';
-import ProductFormFields from "./ProductFormFields";
-import AssetFormFields from "./AssetFormFields";
+import LicensingDetailFormFields from "./LicensingDetailFormFields";
+import ResellerFormFields from "./ResellerFormFields";
 import PaidOutlinedIcon from '@mui/icons-material/PaidOutlined';
 import ProductionQuantityLimitsOutlinedIcon from '@mui/icons-material/ProductionQuantityLimitsOutlined';
 import CategoryOutlinedIcon from '@mui/icons-material/CategoryOutlined';
@@ -36,78 +36,75 @@ const TransactionForm = (props: TransactionFormProps) => {
       fields: TransactionFormFields
     },
     {
-      icon: <ProductionQuantityLimitsOutlinedIcon />,
-      title: 'Product',
-      description: 'Setup your Product data',
-      fields: ProductFormFields
-    },
-    {
-      icon: <CategoryOutlinedIcon />,
-      title: 'Asset',
-      description: 'Setup your Asset data',
-      fields: props.isCreate? AssetFormFields.map(f => ({...f, disabled: true})) : AssetFormFields
-    },
-    {
       icon: <AccountBalanceWalletOutlinedIcon />,
       title: 'Billing Partner Information',
       description: 'Setup your Billing Partner Information',
       fields: PartnerFormFields.BillingPartnerInformationFormFields
     },
+
     {
       icon: <PaidOutlinedIcon />,
       title: 'Shipping Parter Information',
       description: 'Setup your Shipping Partner Information',
       fields: PartnerFormFields.ShippingPartnerInformationFormFields
-    }
+    },
+    {
+      icon: <ProductionQuantityLimitsOutlinedIcon />,
+      title: 'Reseller Information',
+      description: 'Setup your Reseller Information',
+      fields: ResellerFormFields
+    },
+    {
+      icon: <ProductionQuantityLimitsOutlinedIcon />,
+      title: 'Reseller Information',
+      description: 'Setup your Reseller Information',
+      fields: LicensingDetailFormFields
+    },
   ]
   return (
-    <div className="flex flex-col">
-      {
-        FormGroups.map((formGroup, i) => (
-          <Accordion
-            key={i}
-            elevation={0}
-            disableGutters
-            defaultExpanded={i == 0}
-            sx={{
-              // border: '1px solid #d7dde4',
-              borderBottom: '1px solid #d7dde4',
-              '&::before': { display: 'none' },
-            }}
-          >
-            <AccordionSummary
-              expandIcon={<ArrowForwardIosSharpIcon sx={{ fontSize: '0.9rem', color: '#536175' }} />}
-              aria-controls="panel1-content"
-              id={formGroup.title}
+    <div className="flex justify-center">
+      <div className="w-2/3 flex flex-col gap-2">
+        {
+          FormGroups.map((formGroup, i) => (
+            <Accordion
+              key={i}
+              // elevation={0}
+              // disableGutters
+              defaultExpanded={i == 0}
               sx={{
-                py: 1,
-                flexDirection: 'row-reverse',
-                color: '#536175',
-                transitionDuration: '500ms',
-                "&:hover": {
-                  color: "#003133", // Light grey background on hover
-                },
-                '& .MuiAccordionSummary-expandIconWrapper.Mui-expanded': {
-                  transform: 'rotate(90deg)',
-                },
+                borderTop: '2px solid #1f325c',
+                '&::before': { display: 'none' },
+                borderRadius: '0.5rem',
+                boxShadow: '0 10px 15px -3px rgb(0 0 0 / 0.1), 0 4px 6px -4px rgb(0 0 0 / 0.1);'
               }}
             >
-              <div className="pl-4">{formGroup.icon}</div>
-              <div className="pl-2 text-md font-medium">{formGroup.title}</div>
-            </AccordionSummary>
-            <AccordionDetails>
-              <div className="grid grid-cols-3 gap-4 mb-3">
-                <div className="pl-10 text-sm text-[#2f80eb]">
-                  { formGroup.description }
-                </div>
-                <div className="col-span-2">
-                  <GenericForm {...{ ...props, fields: formGroup.fields }} />
-                </div>
-              </div>
-            </AccordionDetails>
-          </Accordion>
-        ))
-      }
+              <AccordionSummary
+                expandIcon={<ArrowForwardIosSharpIcon sx={{ fontSize: '0.9rem', color: '#536175' }} />}
+                aria-controls="panel1-content"
+                id={formGroup.title}
+                sx={{
+                  py: 1,
+                  flexDirection: 'row-reverse',
+                  color: '#536175',
+                  transitionDuration: '500ms',
+                  "&:hover": {
+                    color: "#003133", // Light grey background on hover
+                  },
+                  '& .MuiAccordionSummary-expandIconWrapper.Mui-expanded': {
+                    transform: 'rotate(90deg)',
+                  },
+                }}
+              >
+                <div className="pl-4">{formGroup.icon}</div>
+                <div className="pl-2 text-md font-medium">{formGroup.title}</div>
+              </AccordionSummary>
+              <AccordionDetails>
+                <GenericForm {...{ ...props, fields: formGroup.fields }} />
+              </AccordionDetails>
+            </Accordion>
+          ))
+        }
+      </div>
     </div>
   );
 };
